% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{highcor_tbl}
\alias{highcor_tbl}
\title{Strongest correlations between Cardiometabolic proteins in the UK Biobank used for fallback replacement in case of missingness}
\format{
A data frame with columns:
\describe{
\item{Protein}{character}
\item{Strongest correlated protein}{character}
\item{Correlation coefficient}{numeric}
}
}
\source{
Derived from UK Biobank analyses (also available in Supplementary Table 8)
}
\usage{
highcor_tbl
}
\description{
Strongest correlations between Cardiometabolic proteins in the UK Biobank used for fallback replacement in case of missingness
}
\keyword{datasets}
