% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ProtFIcoef_tbl}
\alias{ProtFIcoef_tbl}
\title{Coefficients for ProtFI}
\format{
A data frame with columns:
\describe{
\item{Protein}{character}
\item{Coefficient}{numeric}
}
}
\source{
Derived for this study (also available in https://github.com/swiergarst/ProtFI/blob/master/output_linear/coefs_frail/selected_models/combine_coefs_frailty_cmb_ffs.csv)
}
\usage{
ProtFIcoef_tbl
}
\description{
Coefficients for ProtFI
}
\keyword{datasets}
