% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ProtMortcoef_tbl}
\alias{ProtMortcoef_tbl}
\title{Coefficients for ProtMort}
\format{
A data frame with columns:
\describe{
\item{Protein}{character}
\item{Coefficient}{numeric}
}
}
\source{
Derived for this study (also available in https://github.com/swiergarst/ProtFI/blob/master/output_linear/coefs_mort/selected_models/combine_coefs_mort_cmb_ffs.csv)
}
\usage{
ProtMortcoef_tbl
}
\description{
Coefficients for ProtMort
}
\keyword{datasets}
